runningService 관련 테스트 코드 작성 및 running duration double형 변환

+++ src/main/java/sprint/server/controller/datatransferobject/response/RunningInfoDTO.java
@@ -7,13 +7,13 @@ import lombok.Getter;
 public class RunningInfoDTO {
 
     private Long runningId;
-    private int duration;
+    private double duration;
     private double distance;
     private String startTime;
     private double energy;
     @Builder
 
-    public RunningInfoDTO(Long runningId, int duration, double distance, String startTime, double energy) {
+    public RunningInfoDTO(Long runningId, double duration, double distance, String startTime, double energy) {
         this.runningId = runningId;
         this.duration = duration;
         this.distance = distance;
